move scene widget to widgets
