+++ src/ndevio/widgets/__init__.py
@@ -1,5 +1,6 @@
 """Widgets for ndevio package."""
 
 from ._plugin_install_widget import PluginInstallerWidget
+from ._scene_widget import DELIMITER, nImageSceneWidget
 
-__all__ = ["PluginInstallerWidget"]
+__all__ = ["PluginInstallerWidget", "nImageSceneWidget", "DELIMITER"]
